backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
openai
pinecone-client
python-dotenv
elevenlabs
